backend/analysis/requirements.txt
fastapi>=0.104.0,<1.0.0
uvicorn[standard]>=0.24.0,<1.0.0
sqlalchemy[asyncio]>=2.0.23,<3.0.0
asyncpg>=0.29.0,<1.0.0
pydantic-settings>=2.1.0,<3.0.0
python-jose[cryptography]>=3.3.0,<4.0.0
passlib[bcrypt]>=1.7.4,<2.0.0
redis>=5.0.0,<6.0.0
httpx>=0.25.0,<1.0.0
pandas>=2.1.0,<3.0.0
numpy>=1.26.0,<2.0.0
influxdb-client>=1.38.0,<2.0.0
aiokafka>=0.10.0,<1.0.0
yfinance>=0.2.31,<1.0.0
python-multipart>=0.0.6,<1.0.0
email-validator>=2.1.0,<3.0.0
